pyproject.toml
[tool.poetry]
name = "words_bot"
version = "0.1.0"
description = "Nasy Telegram Words Bot."
authors = ["Nasy <nasyxx+bot@gmail.com>"]
license = "GPL-3.0+"

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = {version = "^11.1",allows-prereleases = true}
thulac = "^0.2.0"

[tool.black]
line-length=79

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
